--- api/event.go
@@ -1,2 +1,0 @@
-// Package api
-package api
