+++ cmd/api/main.go
@@ -2,81 +2,91 @@ package main
 
 import (
 	"context"
+	"os"
+	"os/signal"
+	"runtime"
+	"syscall"
+
 	"github.com/joho/godotenv"
+	"github.com/kardiachain/kardia-explorer-backend/server/api"
 	"go.uber.org/zap"
 
-	"github.com/kardiachain/kardia-explorer-backend/api"
+	kClient "github.com/kardiachain/go-kaiclient/kardia"
 	"github.com/kardiachain/kardia-explorer-backend/cache"
 	"github.com/kardiachain/kardia-explorer-backend/cfg"
 	"github.com/kardiachain/kardia-explorer-backend/db"
-	"github.com/kardiachain/kardia-explorer-backend/server"
+	"github.com/kardiachain/kardia-explorer-backend/utils"
 )
 
 func main() {
+
 	if err := godotenv.Load(); err != nil {
 		panic(err.Error())
 	}
 
+	runtime.GOMAXPROCS(runtime.NumCPU())
 	serviceCfg, err := cfg.New()
 	if err != nil {
 		panic(err.Error())
 	}
 
-	logger, err := newLogger(serviceCfg)
+	lgr, err := utils.NewLogger(serviceCfg)
 	if err != nil {
 		panic("cannot init logger")
 	}
-	logger.Info("Start API server...")
-
-	defer func() {
-		if err := recover(); err != nil {
-			logger.Error("cannot recover")
-		}
-		if err := logger.Sync(); err != nil {
-			logger.Error("cannot sync log")
+	lgr = lgr.With(zap.String("service_name", "receipts"))
+	lgr.Info("Init receipts services", zap.Any("config", serviceCfg))
+	ctx, cancel := context.WithCancel(context.Background())
+	sigCh := make(chan os.Signal, 1)
+	waitExit := make(chan bool)
+	signal.Notify(sigCh, syscall.SIGINT, syscall.SIGTERM)
+	go func() {
+		for range sigCh {
+			cancel()
+			waitExit <- true
 		}
 	}()
 
-	srvConfig := server.Config{
-		StorageAdapter: db.Adapter(serviceCfg.StorageDriver),
-		StorageURI:     serviceCfg.StorageURI,
-		StorageDB:      serviceCfg.StorageDB,
-		StorageIsFlush: serviceCfg.StorageIsFlush,
+	dbConfig := db.Config{
+		DbAdapter: db.MGO,
+		DbName:    serviceCfg.StorageDB,
+		URL:       serviceCfg.StorageURI,
+		Logger:    lgr,
+		MinConn:   serviceCfg.StorageMinConn,
+		MaxConn:   serviceCfg.StorageMaxConn,
 
-		KardiaURLs:         serviceCfg.KardiaPublicNodes,
-		KardiaTrustedNodes: serviceCfg.KardiaTrustedNodes,
-
-		CacheAdapter:      cache.Adapter(serviceCfg.CacheEngine),
-		CacheURL:          serviceCfg.CacheURL,
-		CacheDB:           serviceCfg.CacheDB,
-		CacheIsFlush:      serviceCfg.CacheIsFlush,
-		BlockBuffer:       serviceCfg.BufferedBlocks,
-		HttpRequestSecret: serviceCfg.HttpRequestSecret,
-
-		Metrics: nil,
-		Logger:  logger,
-
-		AwsAccessKeyId:     serviceCfg.AwsAccessKeyId,
-		AwsSecretAccessKey: serviceCfg.AwsSecretAccessKey,
-		AwsSecretRegion:    serviceCfg.AwsSecretRegion,
-
-		UploaderBucket:     serviceCfg.UploaderBucket,
-		UploaderAcl:        serviceCfg.UploaderAcl,
-		UploaderKey:        serviceCfg.UploaderKey,
-		UploaderPathAvatar: serviceCfg.UploaderPathAvatar,
+		FlushDB: serviceCfg.StorageIsFlush,
 	}
-	srv, err := server.New(srvConfig)
+	dbClient, err := db.NewClient(dbConfig)
 	if err != nil {
-		logger.Panic("cannot create server instance %s", zap.Error(err))
+		lgr.Error("cannot create db conn", zap.Error(err))
+		panic(nil)
 	}
-	ctx := context.Background()
+
+	node, err := kClient.NewNode(serviceCfg.KardiaTrustedNodes[0], lgr)
+	cacheCfg := cache.Config{
+		Adapter:     cache.RedisAdapter,
+		URL:         serviceCfg.CacheURL,
+		DB:          serviceCfg.CacheDB,
+		IsFlush:     serviceCfg.CacheIsFlush,
+		BlockBuffer: serviceCfg.BufferedBlocks,
+		Password:    serviceCfg.CachePassword,
+		Logger:      lgr,
+	}
+	cacheClient, err := cache.New(cacheCfg)
+	if err != nil {
+		lgr.Error("cannot create cache client", zap.Error(err))
+		panic(err)
+	}
+	srv := new(api.Server).
+		SetLogger(lgr).
+		SetStorage(dbClient).
+		SetCache(cacheClient).
+		SetNode(node)
 
 	if serviceCfg.IsReloadBootData {
-		//if err := srv.LoadBootData(ctx); err != nil {
-		//	logger.Panic("cannot load boot data", zap.Error(err))
-		//}
 		if err := srv.LoadBootContracts(ctx); err != nil {
-			logger.Panic("cannot load boot contracts", zap.Error(err))
+			lgr.Panic("cannot load boot contracts", zap.Error(err))
 		}
 	}
