+++ external/cmm.go
@@ -1,0 +1,2 @@
+// Package external
+package external
